requested changes and some code cleaning around my stuff

+++ code/game/objects/items/holosign_creator.dm
@@ -12,49 +12,52 @@
 	throw_speed = 3
 	throw_range = 7
 	item_flags = NOBLUDGEON
-	var/list/signs = list()
+	var/list/signs
 	var/max_signs = 10
 	var/creation_time = 0 //time to create a holosign in deciseconds.
 	var/holosign_type = /obj/structure/holosign/wetsign
 	var/holocreator_busy = FALSE //to prevent placing multiple holo barriers at once
 
-/obj/item/holosign_creator/afterattack(atom/target, mob/user, flag)
+/obj/item/holosign_creator/afterattack(atom/target, mob/user, proximity_flag)
 	. = ..()
-	if(flag)
-		if(!check_allowed_items(target, 1))
+	if(!proximity_flag)
+		return
+	if(!check_allowed_items(target, 1))
+		return
+	var/turf/T = get_turf(target)
+	var/obj/structure/holosign/H = locate(holosign_type) in T
+	if(H)
+		to_chat(user, "<span class='notice'>You use [src] to deactivate [H].</span>")
+		qdel(H)
+		return
+	if(is_blocked_turf(T, TRUE)) //can't put holograms on a tile that has dense stuff
+		return
+	if(holocreator_busy)
+		to_chat(user, "<span class='notice'>[src] is busy creating a hologram.</span>")
+		return
+	if(LAZYLEN(signs) >= max_signs)
+		to_chat(user, "<span class='notice'>[src] is projecting at max capacity!</span>")
+		return
+	playsound(src.loc, 'sound/machines/click.ogg', 20, TRUE)
+	if(creation_time)
+		holocreator_busy = TRUE
+		if(!do_after(user, creation_time, target = target))
+			holocreator_busy = FALSE
 			return
-		var/turf/T = get_turf(target)
-		var/obj/structure/holosign/H = locate(holosign_type) in T
-		if(H)
-			to_chat(user, "<span class='notice'>You use [src] to deactivate [H].</span>")
-			qdel(H)
-		else
-			if(!is_blocked_turf(T, TRUE)) //can't put holograms on a tile that has dense stuff
-				if(holocreator_busy)
-					to_chat(user, "<span class='notice'>[src] is busy creating a hologram.</span>")
-					return
-				if(signs.len < max_signs)
-					playsound(src.loc, 'sound/machines/click.ogg', 20, TRUE)
-					if(creation_time)
-						holocreator_busy = TRUE
-						if(!do_after(user, creation_time, target = target))
-							holocreator_busy = FALSE
-							return
-						holocreator_busy = FALSE
-						if(signs.len >= max_signs)
-							return
-						if(is_blocked_turf(T, TRUE)) //don't try to sneak dense stuff on our tile during the wait.
-							return
-					H = new holosign_type(get_turf(target), src)
-					to_chat(user, "<span class='notice'>You create \a [H] with [src].</span>")
-				else
-					to_chat(user, "<span class='notice'>[src] is projecting at max capacity!</span>")
+		holocreator_busy = FALSE
+		if(LAZYLEN(signs) >= max_signs)
+			return
+		if(is_blocked_turf(T, TRUE)) //don't try to sneak dense stuff on our tile during the wait.
+			return
+	H = new holosign_type(get_turf(target), src)
+	to_chat(user, "<span class='notice'>You create \a [H] with [src].</span>")
+
 
 /obj/item/holosign_creator/attack(mob/living/carbon/human/M, mob/user)
 	return
 
 /obj/item/holosign_creator/attack_self(mob/user)
-	if(signs.len)
+	if(LAZYLEN(signs))
 		for(var/H in signs)
 			qdel(H)
 		to_chat(user, "<span class='notice'>You clear all active holograms.</span>")
@@ -124,7 +127,7 @@
 	if(a.power_equip == FALSE)
 		to_chat(user, "<span class='notice'>[src] can't draw enough power from the grid to create a shield!</span>")
 		return
-	if(signs.len >= max_signs)
+	if(LAZYLEN(signs) >= max_signs)
 		to_chat(user, "<span class='notice'>[src] is projecting at max capacity!</span>")
 		return
 	playsound(loc, 'sound/machines/click.ogg', 20, TRUE)
@@ -155,26 +158,21 @@
 			to_chat(user, "<span class='notice'>You clear all active holograms, and reset your projector to normal.</span>")
 			holosign_type = /obj/structure/holosign/barrier/cyborg
 			creation_time = 5
-			if(signs.len)
+			if(LAZYLEN(signs))
 				for(var/H in signs)
 					qdel(H)
 			shock = 0
 			return
-		else if(R.emagged&&!shock)
+		if(R.emagged&&!shock)
 			to_chat(user, "<span class='warning'>You clear all active holograms, and overload your energy projector!</span>")
 			holosign_type = /obj/structure/holosign/barrier/cyborg/hacked
 			creation_time = 30
-			if(signs.len)
+			if(LAZYLEN(signs))
 				for(var/H in signs)
 					qdel(H)
 			shock = 1
 			return
-		else
-			if(signs.len)
-				for(var/H in signs)
-					qdel(H)
-				to_chat(user, "<span class='notice'>You clear all active holograms.</span>")
-	if(signs.len)
+	if(LAZYLEN(signs))
 		for(var/H in signs)
 			qdel(H)
 		to_chat(user, "<span class='notice'>You clear all active holograms.</span>")
